requirements.txt
Django>=4.0
Pillow>=9.0.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.28.0
